% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.R
\name{.safe_n_workers}
\alias{.safe_n_workers}
\title{Internal: choose a safe number of workers}
\usage{
.safe_n_workers(n_workers, fallback = 1L)
}
\description{
Internal: choose a safe number of workers
}
\keyword{internal}
