% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-docs.R
\name{msclassifr-internal}
\alias{msclassifr-internal}
\alias{alignSpectra_landmark_cpp}
\alias{cpp_match_anchors}
\alias{cpp_warp_mz_piecewise}
\title{Internal helpers and C++ exports}
\description{
These functions are internal building blocks used by SignalProcessingUltra:
}
\details{
\itemize{
\item alignSpectra_landmark_cpp: wrapper that applies landmark matching/warping per spectrum.
\item cpp_match_anchors: Rcpp function for greedy 1:1 matching between spectrum peaks and reference anchors.
\item cpp_warp_mz_piecewise: Rcpp function for piecewise-linear warping of a mass axis.
}

They are not part of the user-facing API.
}
\keyword{internal}
